Rename "wasmer run-unstable" to "wasmer run"
